Update to 1.12.2

+++ src/protocolsupport/protocol/packet/middleimpl/clientbound/play/v_4_5_6_7/KeepAlive.java
@@ -11,7 +11,7 @@ public class KeepAlive extends MiddleKeepAlive {
 	@Override
 	public RecyclableCollection<ClientBoundPacketData> toData() {
 		ClientBoundPacketData serializer = ClientBoundPacketData.create(ClientBoundPacket.PLAY_KEEP_ALIVE_ID, connection.getVersion());
-		serializer.writeInt(keepAliveId);
+		serializer.writeInt((int) keepAliveId);
 		return RecyclableSingletonList.create(serializer);
 	}
